Promote imported FANTOIR version
Summary:
When a new FANTOIR file has been imported to a new table,
declare in the configuration the new table to use.

For convenience, a fantoir view is also provided, as an alias
to the promoted table.

When other tables will need a foreign key to the fantoir table,
this command will also be responsible to update them.

This strategy allows to transparently update the datasource
while allowing consumers still to use it pending update.

Ref T1750.

Test Plan:
  - run `fantoir-datasource promote fantoir_202210`
  - check view `fantoir` and table `fantoir_config` are created
  - run it again to check
 idempotency

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1750

Differential Revision: https://devcentral.nasqueron.org/D2733