Search the imported FANTOIR database
Summary:
When data is imported from Wikidata, the code direction can be missing.
When it's can be computed, like for departments actually using them,
we can query the database to find the code.

Implement a query service, and a CLI front-end.

The CLI front-end is minimalist, mainly to let test if data is correctly
imported after an import task before promotion.

Ref T1750, T1751.

Test Plan: `./fantoir-datasource query ...`

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1751, T1750

Differential Revision: https://devcentral.nasqueron.org/D2734